+++ android/app/src/main/java/com/wyrazowo/FSModuleManager.kt
@@ -1,0 +1,47 @@
+package com.wyrazowo
+import android.app.Activity
+import android.content.Intent
+import android.os.Bundle
+import com.facebook.react.bridge.BaseActivityEventListener
+import com.facebook.react.bridge.ReactApplicationContext
+import com.facebook.react.bridge.ReactContextBaseJavaModule
+import com.facebook.react.bridge.ReactMethod
+import com.facebook.react.modules.core.DeviceEventManagerModule
+
+class FSModuleManager(reactContext: ReactApplicationContext): ReactContextBaseJavaModule(reactContext) {
+    override fun getName(): String {
+      return "FSModule"
+    }
+
+    private val listener: BaseActivityEventListener = object : BaseActivityEventListener() {
+      override fun onActivityResult(activity: Activity, requestCode: Int, resultCode: Int, data: Intent?) {
+        super.onActivityResult(activity, requestCode, resultCode, data)
+        if (requestCode == 0) {
+          if (resultCode == Activity.RESULT_OK) {
+            reactApplicationContext
+              .getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter::class.java)
+              .emit("readSearchHistory", data!!.getStringExtra("readData"))
+          }
+        }
+      }
+    }
+
+    init {
+      reactApplicationContext?.addActivityEventListener(listener)
+    }
+
+    @ReactMethod
+    fun saveSearchHistory(searchHistory: String) {
+      val createFileIntent = Intent(reactApplicationContext, FSActivity::class.java)
+      createFileIntent.putExtra("dataToSave", searchHistory)
+      createFileIntent.flags = Intent.FLAG_ACTIVITY_NEW_TASK
+      reactApplicationContext.startActivity(createFileIntent)
+    }
+
+    @ReactMethod
+    fun readSearchHistory() {
+      val readFileIntent = Intent(reactApplicationContext, FSActivity::class.java)
+      readFileIntent.flags = Intent.FLAG_ACTIVITY_MULTIPLE_TASK
+      reactApplicationContext.startActivityForResult(readFileIntent, 0, Bundle.EMPTY)
+    }
+}