+++ ios/RCTDBModule.swift
@@ -1,8 +1,0 @@
-//
-//  RCTDBModule.swift
-//  Wyrazowo
-//
-//  Created by Adam Kruczek on 03/04/2023.
-//
-
-import Foundation