+++ src/core/header/header.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { Platform, StatusBar } from 'react-native'
+import { StatusBar } from 'react-native'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 import { useNavigation } from '@react-navigation/native'
 import { genericLightShadow, genericTextShadow } from '@core/shadow/shadow.constants'
@@ -7,6 +7,7 @@ import { COLOR } from '@core/colors/colors.constants'
 import { useLocalize } from '@core/hooks/use-localize.hook'
 import { isPlatform } from '@core/is-platform/is-platform'
 import { HeaderSideContentConfig, HeaderType } from './header.models'
+import { useHeaderTextSize } from './hooks/use-header-text-size.hook'
 import {
   BackButtonContainer, BackButtonIcon, HeaderContainer, HeaderLeftButtonContainer, HeaderLeftButtonIndicator,
   HeaderLeftIcon, HeaderRightButtonContainer, HeaderRightButtonIndicator, HeaderRightIcon, HeaderStatusBar, HeaderText,
@@ -24,6 +25,7 @@ export const Header = ({ type, title, backButton, rightContentConfig, leftConten
   const navigation = useNavigation()
   const { top: topInset } = useSafeAreaInsets()
   const localize = useLocalize()
+  const { onHeaderTextLayout, headerTextSize } = useHeaderTextSize()
 
   const color: COLOR = {
     dashboard: COLOR.FIRE_BRICK,
@@ -49,6 +51,8 @@ export const Header = ({ type, title, backButton, rightContentConfig, leftConten
 
         <HeaderText
           style={genericTextShadow}
+          headerTextSize={headerTextSize}
+          onLayout={onHeaderTextLayout}
           children={` ${title ?? localize()[type]} `}
         />